fix: ajustando e finalizando a feature de produto e categoria

+++ src/category/category.service.ts
@@ -1,14 +1,28 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
-import { PrismaService } from 'src/prisma/prisma.service';
-import { CreateCategoryDto } from './dto/create-category.dto';
-import { UpdateCategoryDto } from './dto/update-category.dto';
+import {
+  BadRequestException,
+  ConflictException,
+  Injectable,
+  NotFoundException,
+} from "@nestjs/common";
+import { PrismaService } from "src/prisma/prisma.service";
+import { CreateCategoryDto } from "./dto/create-category.dto";
+import { UpdateCategoryDto } from "./dto/update-category.dto";
 
 @Injectable()
 export class CategoryService {
   constructor(private prisma: PrismaService) {}
 
-  create(dto: CreateCategoryDto) {
-    return this.prisma.category.create({ data: dto });
+  async create(dto: CreateCategoryDto) {
+    const exists = await this.prisma.category.findFirst({
+      where: { name: { equals: dto.name, mode: "insensitive" } },
+    });
+
+    if (exists) throw new ConflictException("Categoria já existe");
+    return this.prisma.category.create({
+      data: {
+        name: dto.name.trim(),
+      },
+    });
   }
 
   findAll() {
@@ -16,6 +30,7 @@ export class CategoryService {
       include: {
         produtos: true,
       },
+      orderBy: { name: "asc" },
     });
   }
 
@@ -29,13 +44,24 @@ export class CategoryService {
       },
     });
 
-    if (!category) throw new NotFoundException('Categoria não encontrada');
+    if (!category) throw new NotFoundException("Categoria não encontrada");
 
     return category;
   }
 
   async update(id: string, dto: UpdateCategoryDto) {
     await this.findOne(id);
+
+    if (dto.name) {
+      const exists = await this.prisma.category.findFirst({
+        where: {
+          name: { equals: dto.name.trim(), mode: "insensitive" },
+          NOT: { id },
+        },
+      });
+
+      if (exists) throw new ConflictException("Categoria já existe");
+    }
 
     return this.prisma.category.update({
       where: {
@@ -47,6 +73,15 @@ export class CategoryService {
 
   async remove(id: string) {
     await this.findOne(id);
+
+    const hasProducts = await this.prisma.product.count({
+      where: { categoryId: id },
+    });
+
+    if (hasProducts > 0)
+      throw new BadRequestException(
+        "Não é possível deletar uma categoria que possui produtos associados"
+      );
     return this.prisma.category.delete({ where: { id } });
   }
 }
